refactor(api): implement modular news extractor pattern and centralize utilities
